requirements.txt
pydantic
python-dotenv
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-groq
langchain-core
langchain-community>=0.0.20
langchain_google_genai
langchain_huggingface
langchain-pinecone
langchain-mongodb>=0.1.0
defusedxml
duckduckgo-search
pypdf
bs4
arxiv
pymupdf
wikipedia
faiss-cpu
pandas>=2.0.0
openai>=1.12.0
pymongo>=4.6.0
pillow
pdfplumber
tqdm
camelot-py[cv]>=0.11.0
marimo[recommended]
python-docx>=1.1.0
streamlit>=1.28.0
rank-bm25>=0.2.2
scikit-learn>=1.3.0
numpy>=1.24.0